+++ tests/test_x86.c
@@ -19,7 +19,7 @@ struct platform {
 	cs_opt_value opt_value;
 };
 
-static void print_string_hex(char *comment, unsigned char *str, int len)
+static void print_string_hex(char *comment, unsigned char *str, size_t len)
 {
 	unsigned char *c;
